feat: monkey patch for windows kill

+++ src/deno/BrowserRunner.ts
@@ -125,7 +125,11 @@ export class BrowserRunner {
     // `proc.kill` would error, as the `pid` to-be-killed can not be found.
     if (this.proc && this.proc.pid && !this._closed) {
       try {
-        this.proc.kill(9);
+        if (Deno.build.os === "windows") {
+          Deno.run({ cmd: ["taskkill", "/f", "/pid", `${this.proc.pid}`] });
+        } else {
+          this.proc.kill(9);
+        }
       } catch (error) {
         throw new Error(
           `${PROCESS_ERROR_EXPLANATION}\nError cause: ${error.stack}`,
